refactor: fix design principle violations and extract service layer
